+++ MCP_ADC.cpp
@@ -150,20 +150,12 @@ int16_t MCP_ADC::readADC(uint8_t channel, bool single)
   if (_hwSPI)
   {
     _mySPI->beginTransaction(_spi_settings);
-    //  _mySPI->transfer(data, bytes);  //  TODO test with HW.
-    for (uint8_t b = 0; b < bytes; b++)
-    {
-      data[b] = _mySPI->transfer(data[b]);
-    }
+    _mySPI->transfer(data, bytes);  //  0.5.3
     _mySPI->endTransaction();
   }
   else  //  Software SPI
   {
-    //  swSPI_transfer(data, bytes);  //  experimental
-    for (uint8_t b = 0; b < bytes; b++)
-    {
-      data[b] = swSPI_transfer(data[b]);
-    }
+    swSPI_transfer(data, bytes);  //  0.5.3
   }
   digitalWrite(_select, HIGH);
 
@@ -199,19 +191,11 @@ void MCP_ADC::readADCMultiple(uint8_t channels[], uint8_t numChannels, int16_t r
 
     if (_hwSPI)
     {
-      //  _mySPI->transfer(data, bytes);  //  experimental
-      for (uint8_t b = 0; b < bytes; b++)
-      {
-        data[b] = _mySPI->transfer(data[b]);
-      }
+      _mySPI->transfer(data, bytes);  //  0.5.3
     }
     else
     {
-    // swSPI_transfer(data, bytes);  //  experimental
-    for (uint8_t b = 0; b < bytes; b++)
-    {
-      data[b] = swSPI_transfer(data[b]);
-    }
+      swSPI_transfer(data, bytes);  //  0.5.3
     }
 
     if (bytes == 2) {
@@ -235,46 +219,32 @@ void MCP_ADC::readADCMultiple(uint8_t channels[], uint8_t numChannels, int16_t r
 }
 
 
-//  MSBFIRST
-uint8_t  MCP_ADC::swSPI_transfer(uint8_t val)
+//  0.5.3
+//  MSB FIRST
+void  MCP_ADC::swSPI_transfer(uint8_t * data, uint8_t bytes)
 {
   uint8_t clk = _clock;
   uint8_t dao = _dataOut;
   uint8_t dai = _dataIn;
 
-  uint8_t rv = 0;
-  for (uint8_t mask = 0x80; mask; mask >>= 1)
+  for (uint8_t i = 0; i < bytes; i++)
   {
-    digitalWrite(dao, (val & mask));
-    digitalWrite(clk, HIGH);
-    if (digitalRead(dai) == HIGH) rv |= mask;
-    digitalWrite(clk, LOW);
+    uint8_t rv = 0;
+    uint8_t val = data[i];
+    for (uint8_t mask = 0x80; mask; mask >>= 1)
+    {
+      digitalWrite(dao, (val & mask));
+      digitalWrite(clk, HIGH);
+//  reduce fast processors (32MHz) to max 1 MHz
+#if F_CPU >= 32000000UL
+    delayMicroseconds(1);
+#endif
+      if (digitalRead(dai) == HIGH) rv |= mask;
+      digitalWrite(clk, LOW);
+    }
+    data[i] = rv;
   }
-  return rv;
 }
-
-//  EXPERIMENTAL
-// void  MCP_ADC::swSPI_transfer(uint8_t * data, uint8_t bytes)
-// {
-  // uint8_t clk = _clock;
-  // uint8_t dao = _dataOut;
-  // uint8_t dai = _dataIn;
-
-  // for (uint8_t i = 0; i < bytes; i++)
-  // {
-    // uint8_t rv = 0;
-    // uint8_t val = data[i];
-    // for (uint8_t mask = 0x80; mask; mask >>= 1)
-    // {
-      // digitalWrite(dao, (val & mask));
-      // digitalWrite(clk, HIGH);
-      // if (digitalRead(dai) == HIGH) rv |= mask;
-      // digitalWrite(clk, LOW);
-    // }
-    // data[i] = rv;
-  // }
-  // // return rv;
-// }
 
 
 /////////////////////////////////////////////////////////////////////////////