+++ MCP_ADC.h
@@ -69,7 +69,7 @@ protected:
   uint8_t  _channels;
   int16_t  _maxValue;
   //  1 MHz is a safe value (datasheet)
-  //  in a test 4 MHz worked => #26 advices 2 MHz at most 
+  //  in a test 4 MHz worked => #26 advices 2 MHz at most
   uint32_t _SPIspeed = 1000000;
 
   //  derived classes must implement buildRequest() function.
@@ -77,10 +77,8 @@ protected:
 
   int16_t  readADC(uint8_t channel, bool single);
   void     readADCMultiple(uint8_t channels[], uint8_t numChannels, int16_t readings[]);
-
-  uint8_t  swSPI_transfer(uint8_t d);
-  //  experimental
-  //  void     swSPI_transfer(uint8_t * data, uint8_t bytes);
+  //  0.5.3
+  void     swSPI_transfer(uint8_t * data, uint8_t bytes);
 
   __SPI_CLASS__ * _mySPI;
   SPISettings   _spi_settings;